pyproject.toml
[project]
name = "upstage-infra-project"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=46.0.3",
    "pymysql>=1.1.2",
    "sqlalchemy>=2.0.44",
    "python-dotenv>=1.2.1",
    "fastapi>=0.128.0",
    "uvicorn>=0.40.0",
    "mysql-connector-python>=9.5.0",
    "loguru>=0.7.3",
    "httpx>=0.28.1",
]
